mqtt/requirements.txt
kafka-python
paho-mqtt
faust
typing-extensions
numpy
cassandra-driver
mysql-connector-python
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-proto-grpc